Moved functionality related to addresses to a separate AddressService
