+++ core/src/test/java/greencity/controller/AddressControllerTest.java
@@ -9,7 +9,6 @@ import greencity.dto.CreateAddressRequestDto;
 import greencity.dto.location.api.DistrictDto;
 import greencity.dto.order.OrderAddressDtoRequest;
 import greencity.service.ubs.AddressService;
-import greencity.service.ubs.UBSClientService;
 import greencity.service.ubs.UBSManagementService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -47,9 +46,6 @@ class AddressControllerTest {
     private MockMvc mockMvc;
 
     @Mock
-    private UBSClientService ubsClientService;
-
-    @Mock
     private UserRemoteClient userRemoteClient;
 
     @Mock
@@ -80,7 +76,7 @@ class AddressControllerTest {
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk());
 
-        verify(ubsClientService).findAllAddressesForCurrentOrder(anyString());
+        verify(addressService).findAllAddressesForCurrentOrder(anyString());
     }
 
     @Test
@@ -97,7 +93,7 @@ class AddressControllerTest {
             .principal(principal)
             .contentType(MediaType.APPLICATION_JSON)).andExpect(status().isCreated());
 
-        verify(ubsClientService).saveCurrentAddressForOrder(any(), eq("35467585763t4sfgchjfuyetf"));
+        verify(addressService).saveCurrentAddressForOrder(any(), eq("35467585763t4sfgchjfuyetf"));
     }
 
     @Test
@@ -114,7 +110,7 @@ class AddressControllerTest {
             .principal(principal)
             .contentType(MediaType.APPLICATION_JSON)).andExpect(status().isOk());
 
-        verify(ubsClientService).updateCurrentAddressForOrder(any(), eq("35467585763t4sfgchjfuyetf"));
+        verify(addressService).updateCurrentAddressForOrder(any(), eq("35467585763t4sfgchjfuyetf"));
     }
 
     @Test
@@ -135,7 +131,7 @@ class AddressControllerTest {
             .principal(principal))
             .andExpect(status().isOk());
 
-        verify(ubsClientService).makeAddressActual(addressId, uuid);
+        verify(addressService).makeAddressActual(addressId, uuid);
     }
 
     @Test
@@ -143,7 +139,7 @@ class AddressControllerTest {
         String region = "Львівська";
         String city = "Львів";
         List<DistrictDto> mockLocationDtoList = new ArrayList<>();
-        when(ubsClientService.getAllDistricts(region, city)).thenReturn(mockLocationDtoList);
+        when(addressService.getAllDistricts(region, city)).thenReturn(mockLocationDtoList);
         mockMvc.perform(get(ubsLink + "/get-all-districts")
             .param("region", region)
             .param("city", city)
@@ -151,7 +147,7 @@ class AddressControllerTest {
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk());
 
-        verify(ubsClientService).getAllDistricts(region, city);
+        verify(addressService).getAllDistricts(region, city);
     }
 
     @Test
@@ -172,7 +168,7 @@ class AddressControllerTest {
             .principal(principal))
             .andExpect(status().isOk());
 
-        verify(managementService).addressUpdate(any(), eq(principal.getName()));
+        verify(addressService).addressUpdate(any(), eq(principal.getName()));
     }
 
     @Test
