+++ service-api/src/main/java/greencity/service/ubs/UBSManagementService.java
@@ -1,6 +1,5 @@
 package greencity.service.ubs;
 
-import greencity.dto.address.UpdateAddressDto;
 import greencity.dto.bag.AdditionalBagInfoDto;
 import greencity.dto.certificate.CertificateDtoForSearching;
 import greencity.dto.employee.EmployeePositionDtoRequest;
@@ -13,15 +12,12 @@ import greencity.dto.order.EcoNumberDto;
 import greencity.dto.order.ExportDetailsDto;
 import greencity.dto.order.ExportDetailsDtoUpdate;
 import greencity.dto.order.NotTakenOrderReasonDto;
-import greencity.dto.order.OrderAddressDtoResponse;
-import greencity.dto.order.OrderAddressExportDetailsDtoUpdate;
 import greencity.dto.order.OrderCancellationReasonDto;
 import greencity.dto.order.OrderDetailInfoDto;
 import greencity.dto.order.OrderDetailStatusDto;
 import greencity.dto.order.OrderDetailStatusRequestDto;
 import greencity.dto.order.OrderInfoDto;
 import greencity.dto.order.OrderStatusPageDto;
-import greencity.dto.order.ReadAddressByOrderDto;
 import greencity.dto.order.UpdateAllOrderPageDto;
 import greencity.dto.order.UpdateOrderPageAdminDto;
 import greencity.dto.pageble.PageableDto;
@@ -29,9 +25,10 @@ import greencity.dto.user.AddingPointsToUserDto;
 import greencity.dto.violation.ViolationsInfoDto;
 import greencity.entity.order.Order;
 import greencity.enums.SortingOrder;
+
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
+
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -43,7 +40,7 @@ public interface UBSManagementService {
      * @author Nazar Struk
      */
     PageableDto<CertificateDtoForSearching> getAllCertificates(Pageable page, String columnName,
-        SortingOrder sortingOrder);
+                                                               SortingOrder sortingOrder);
 
     /**
      * Method add some points to UserUBS by email.
@@ -56,31 +53,10 @@ public interface UBSManagementService {
      * Method returns all users violations.
      *
      * @return {@link ViolationsInfoDto} count of Users violations with order id
-     *         descriptions.
+     * descriptions.
      * @author Nazar Struk
      */
     ViolationsInfoDto getAllUserViolations(String email);
-
-    /**
-     * Method that read user address by order id.
-     *
-     * @param orderId of {@link Long} order id;
-     * @return {@link ReadAddressByOrderDto} that contains one address;
-     * @author Mahdziak Orest
-     */
-    ReadAddressByOrderDto getAddressByOrderId(Long orderId);
-
-    /**
-     * Method that update address.
-     *
-     * @param dtoUpdate of {@link OrderAddressExportDetailsDtoUpdate} order id.
-     * @param order     {@link Order}.
-     * @param email     {@link String}.
-     * @return {@link OrderAddressDtoResponse} that contains address.
-     * @author Mahdziak Orest
-     */
-    Optional<OrderAddressDtoResponse> updateAddress(OrderAddressExportDetailsDtoUpdate dtoUpdate, Order order,
-        String email);
 
     /**
      * Method for getting order detail by language and order id.
@@ -95,7 +71,7 @@ public interface UBSManagementService {
      * @author Mahdziak Orest
      */
     void setOrderDetail(Order order,
-        Map<Integer, Integer> confirmed, Map<Integer, Integer> exported, String email);
+                        Map<Integer, Integer> confirmed, Map<Integer, Integer> exported, String email);
 
     /**
      * Method that count sum order.
@@ -137,7 +113,6 @@ public interface UBSManagementService {
      * Method that returns total number of orders.
      *
      * @return {@link Long}.
-     *
      * @author Chernenko Vitaliy
      */
     OrderCountDto getTotalNumberOfOrders();
@@ -201,7 +176,6 @@ public interface UBSManagementService {
      *
      * @param adminCommentDto {@link AdminCommentDto}.
      * @param email           {@link String}.
-     *
      * @author Yuriy Bahlay.
      */
     void saveAdminCommentToOrder(AdminCommentDto adminCommentDto, String email);
@@ -212,7 +186,6 @@ public interface UBSManagementService {
      * @param ecoNumberDto {@link EcoNumberDto}.
      * @param orderId      {@link Long}.
      * @param email        {@link String}.
-     *
      * @author Yuriy Bahlay.
      */
     void updateEcoNumberForOrderById(EcoNumberDto ecoNumberDto, Long orderId, String email);
@@ -223,7 +196,6 @@ public interface UBSManagementService {
      * @param ecoNumberDto {@link EcoNumberDto}.
      * @param order        {@link Order}.
      * @param email        {@link String}.
-     *
      * @author Yuriy Bahlay.
      */
     void updateEcoNumberForOrder(EcoNumberDto ecoNumberDto, Order order, String email);
@@ -236,12 +208,11 @@ public interface UBSManagementService {
      * @param language                {@link String}.
      * @param email                   {@link String}.
      * @param images                  {@link MultipartFile}.
-     *
      * @author Anton Bondar.
      */
     BigOrderTableDTO updateOrderAdminPageInfoAndSaveReason(Long orderId,
-        UpdateOrderPageAdminDto updateOrderPageAdminDto,
-        String language, String email, MultipartFile[] images);
+                                                           UpdateOrderPageAdminDto updateOrderPageAdminDto,
+                                                           String language, String email, MultipartFile[] images);
 
     /**
      * This is method which is updates admin page info for order.
@@ -249,18 +220,16 @@ public interface UBSManagementService {
      * @param updateOrderPageAdminDto {@link UpdateOrderPageAdminDto}.
      * @param order                   {@link Order}.
      * @param email                   {@link String}.
-     *
      * @author Yuriy Bahlay.
      */
     void updateOrderAdminPageInfo(UpdateOrderPageAdminDto updateOrderPageAdminDto, Order order, String lang,
-        String email);
+                                  String email);
 
     /**
      * This is method which is updates admin page info for all order.
      *
      * @param updateAllOrderPageDto {@link UpdateOrderPageAdminDto}.
      * @param email                 {@link String} currentUser.
-     *
      * @author Max Boiarchuk.
      */
     void updateAllOrderAdminPageInfo(UpdateAllOrderPageDto updateAllOrderPageDto, String email, String lang);
@@ -270,7 +239,6 @@ public interface UBSManagementService {
      *
      * @param orderId {@link Long}.
      * @param email   {@link String}.
-     *
      * @author Hlazova Nataliia.
      */
     Boolean checkEmployeeForOrder(Long orderId, String email);
@@ -288,7 +256,6 @@ public interface UBSManagementService {
      *
      * @param orderId {@link Long}.
      * @return {@link OrderCancellationReasonDto}
-     *
      * @author Kharchenko Volodymyr.
      */
     OrderCancellationReasonDto getOrderCancellationReason(Long orderId);
@@ -298,7 +265,6 @@ public interface UBSManagementService {
      *
      * @param orderId {@link Long}.
      * @return {@link NotTakenOrderReasonDto}.
-     *
      * @author Kharchenko Volodymyr.
      */
     NotTakenOrderReasonDto getNotTakenOrderReason(Long orderId);
@@ -309,21 +275,8 @@ public interface UBSManagementService {
      *
      * @param orderId {@link Long} the ID of the order.
      * @return {@link Boolean}
-     *
      * @author Volodymyr Lukovskyi
      */
     Boolean checkIfOrderStatusIsFormedToCanceled(Long orderId);
 
-    /**
-     * Method updates order address. This method updates order address. It takes
-     * {@link UpdateAddressDto} as an argument and updates the order address in the
-     * database. It also checks if the address exists and if the user is authorized
-     * to update the address.
-     *
-     * @param addressDto {@link UpdateAddressDto}
-     * @param email      {@link String} the user's email
-     *
-     * @author Kizerov Dmytro
-     */
-    void addressUpdate(UpdateAddressDto addressDto, String email);
 }
